+++ concord-ui/src/components/searchbar.css
